new app 'products' was created, refactoring in 'mainapp' has been made
