--- Geekshop/views.py
@@ -1,9 +1,0 @@
-from django.shortcuts import render
-
-
-def home(request):
-    return render(request, 'index.html')
-
-
-def contacts(request):
-    return render(request, 'contact.html')
