+++ mainapp/static/mainapp/fonts/font-awesome/css/font-awesome.css
