+++ mainapp/templates/mainapp/contact.html
@@ -3,8 +3,8 @@
 <head>
 	<meta charset="utf-8">
 	<title>Контакты</title>
-	<link rel="stylesheet" type="text/css" href="/static/geekshop/style.css">
-	<link rel="stylesheet" href="/static/geekshop/fonts/font-awesome/css/font-awesome.css">
+	<link rel="stylesheet" type="text/css" href="/static/mainapp/style.css">
+	<link rel="stylesheet" href="/static/mainapp/fonts/font-awesome/css/font-awesome.css">
 </head>
 <body>
 	<div class="container">
