+++ mainapp/templates/mainapp/index.html
@@ -3,8 +3,8 @@
 <head>
 	<meta charset="utf-8">
 	<title>Магазин</title>
-	<link rel="stylesheet" type="text/css" href="/static/geekshop/style.css">
-	<link rel="stylesheet" href="/static/geekshop/fonts/font-awesome/css/font-awesome.css">
+	<link rel="stylesheet" type="text/css" href="/static/mainapp/style.css">
+	<link rel="stylesheet" href="/static/mainapp/fonts/font-awesome/css/font-awesome.css">
 </head>
 <body>
 	<div class="container">
@@ -22,7 +22,7 @@
 				<h3 class="green">тенденции</h3>
 				<h3 class="big"> удобные стулья</h3>
 				<p>Новый уровень комфорта. Отличные характеристики.</p>
-				<a href="#" class="button-order">заказать <img src="/static/geekshop/img/arrow.png" alt=""></a>
+				<a href="#" class="button-order">заказать <img src="/static/mainapp/img/arrow.png" alt=""></a>
 			</div>	
 			<div class="navi">
 				<a href="#"></a><a href="#"></a><a href="#"></a>
@@ -33,32 +33,32 @@
 			<div class="featured-menu">
 					<a class="active" href="#">Популярные</a><a href="#">Новинки</a>
 			</div>
-			<a class="explore" href="#">показать все <img src="/static/geekshop/img/arrow.png" alt=""></a>
+			<a class="explore" href="#">показать все <img src="/static/mainapp/img/arrow.png" alt=""></a>
 			<div class="clearfix">
 			</div>
 			<div class="tab-content">
 				<div class="block">
-					<a href="#"><img src="/static/geekshop/img/product-1.jpg" alt="">
+					<a href="#"><img src="/static/mainapp/img/product-1.jpg" alt="">
 						<div class="text">
-							<img src="/static/geekshop/img/icon-hover.png" alt="img">
+							<img src="/static/mainapp/img/icon-hover.png" alt="img">
 							<h4>Отличный стул</h4>
 							<p>Расположитесь комфортно. </p>
 						</div> 
 					</a>
 				</div>
 				<div class="block">
-					<a href="#"><img src="/static/geekshop/img/product-3.jpg" alt="">
+					<a href="#"><img src="/static/mainapp/img/product-3.jpg" alt="">
 						<div class="text">
-							<img src="/static/geekshop/img/icon-hover.png" alt="img">
+							<img src="/static/mainapp/img/icon-hover.png" alt="img">
 							<h4>Супер стул</h4>
 							<p>Расположитесь комфортно. </p>
 						</div>
 					</a>
 				</div>
 				<div class="block">
-					<a href="#"><img src="/static/geekshop/img/product-2.jpg" alt="">
+					<a href="#"><img src="/static/mainapp/img/product-2.jpg" alt="">
 						<div class="text">
-							<img src="/static/geekshop/img/icon-hover.png" alt="">
+							<img src="/static/mainapp/img/icon-hover.png" alt="">
 							<h4>Стул повышенного качества</h4>
 							<p>Не оторваться. </p>
 						</div>
