+++ mainapp/views.py
@@ -1,5 +1,9 @@
 from django.shortcuts import render
 
 
-def products(request):
-    return render(request, 'products.html')
+def home(request):
+    return render(request, 'mainapp/index.html')
+
+
+def contacts(request):
+    return render(request, 'mainapp/contact.html')
