+++ products/templates/products/products.html
@@ -3,9 +3,9 @@
 <head>
 	<meta charset="utf-8">
 	<title>Каталог</title>
-	<link rel="stylesheet" href="/static/geekshop/style.css">
+	<link rel="stylesheet" href="/static/mainapp/style.css">
 	<link rel="stylesheet" href="/static/products/products_style.css">
-	<link rel="stylesheet" href="/static/geekshop/fonts/font-awesome/css/font-awesome.css">
+	<link rel="stylesheet" href="/static/mainapp/fonts/font-awesome/css/font-awesome.css">
 </head>
 <body>
 	<div class="container">
@@ -37,28 +37,28 @@
 			<div class="details-products">
 				<div class="details-slider">
 					<div class="slider-product">
-						<img src="/static/geekshop/img/slider1.jpg">
+						<img src="/static/mainapp/img/slider1.jpg">
 					</div>
 					<div class="slider-control">
 						<div class="block">
-							<a href="#"><img src="/static/geekshop/img/controll.jpg" alt="">
+							<a href="#"><img src="/static/mainapp/img/controll.jpg" alt="">
 								<div class="text"></div>
 							</a>
 						</div>
 						<div class="block">
-							<a href="#"><img src="/static/geekshop/img/controll1.jpg" alt="">
+							<a href="#"><img src="/static/mainapp/img/controll1.jpg" alt="">
 								<div class="text"></div>
 							</a>
 						</div>
 						<div class="block">
-							<a href="#"><img src="/static/geekshop/img/controll2.jpg" alt="">
+							<a href="#"><img src="/static/mainapp/img/controll2.jpg" alt="">
 								<div class="text"></div>
 							</a>
 						</div>
 					</div>
 				</div>
 
-				<div class="description">
+				<div class="description description__upd">
 					<h3 class="big-bold__upd">Отличный стул</h3>
 					<h3 class="red">горячее предложение</h3>
 					<p class="price"> 2585.9 <span>руб</span></p>
@@ -79,31 +79,31 @@
 			<div class="related">
 				<div class="title clearfix">
 					<h2>Похожие продукты</h2>
-					<a class="explore" href="#">показать все <img src="/static/geekshop/img/arrow.png" alt=""></a>
+					<a class="explore" href="#">показать все <img src="/static/mainapp/img/arrow.png" alt=""></a>
 				</div>
 				<div class="related-products clearfix">
 					<div class="block">
-						<a href="#"><img src="/static/geekshop/img/product-11.jpg" alt="">
+						<a href="#"><img src="/static/mainapp/img/product-11.jpg" alt="">
 							<div class="text">
-								<img src="/static/geekshop/img/icon-hover.png" alt="img">
+								<img src="/static/mainapp/img/icon-hover.png" alt="img">
 								<h4>Стул повышенного качества</h4>
 								<p>Не оторваться. </p>
 							</div>
 						</a>
 					</div>
 					<div class="block">
-						<a href="#"><img src="/static/geekshop/img/product-21.jpg" alt="">
+						<a href="#"><img src="/static/mainapp/img/product-21.jpg" alt="">
 							<div class="text">
-								<img src="/static/geekshop/img/icon-hover.png" alt="img">
+								<img src="/static/mainapp/img/icon-hover.png" alt="img">
 								<h4>Стул повышенного качества</h4>
 								<p>Не оторваться. </p>
 							</div>
 						</a>
 					</div>
 					<div class="block">
-						<a href="#"><img src="/static/geekshop/img/product-31.jpg" alt="">
+						<a href="#"><img src="/static/mainapp/img/product-31.jpg" alt="">
 							<div class="text">
-								<img src="/static/geekshop/img/icon-hover.png" alt="img">
+								<img src="/static/mainapp/img/icon-hover.png" alt="img">
 								<h4>Стул повышенного качества</h4>
 								<p>Не оторваться. </p>
 							</div>
