+++ products/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import products
+
+urlpatterns = [
+    path('', products, name='products'),
+]
