+++ products/views.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def products(request):
+    return render(request, 'products/products.html')
